app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0

opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0

python-json-logger==2.0.7
